fuck fuck fuck v1.1

+++ main-service/src/main/java/ru/practicum/ewm/mainservice/service/privateApi/PrivateRequestsServiceImpl.java
@@ -48,7 +48,7 @@ public class PrivateRequestsServiceImpl implements PrivateRequestsService {
         if (!participationRequest.isEmpty()) {
             throw new DataIntegrityViolationException("Нельзя добавить повторный запрос!");
         }
-        if (event.getInitiator().getId() == userId) {
+        if (event.getInitiator().getId().longValue() == userId.longValue()) {
             throw new DataIntegrityViolationException("Инициатор события не может добавить запрос на участие в своём событии!");
         }
         if (!event.getState().equals(EventState.PUBLISHED)) {
